requirements.txt
typing-inspect==0.8.0
typing_extensions==4.8.0
openai==0.27.8
langchain==0.0.286
chroma==0.2.0
chromadb==0.4.5
pydantic==1.10.11 
faiss-cpu
tiktoken
PyPDF
streamlit
streamlit-chat
streamlit-extras
streamlit_option_menu
msal-streamlit-authentication==1.0.9
pandasai==0.8.1
neo4j
streamlit_js_eval==0.1.5
psycopg2-binary
pgvector
docx2txt
llama-index==0.8.29
sqlalchemy[asyncio]
asyncpg
nltk
boto3
streamlit-feedback
scikit-learn==1.3.2
streamlit-audiorec
playsound==1.2.2
openpyxl
sounddevice
amazon_transcribe